refactor(engines): split LaTeX-log renderer from parser
Move LatexLogRenderer (Rich-console presentation) out of the 720-line log.py
into a sibling log_render module (log down to ~500 lines), separating log
parsing from presentation. stream_latexmk_output imports the renderer lazily
to avoid a parser<->renderer import cycle; shared message types and highlight
patterns stay in log. Package __init__ re-exports LatexLogRenderer from the
new module; test_latex_log (path-loads log.py for the parser) is unaffected.

All 849 tests green, ruff clean.
